mgr/dashboard: Add generic Select component
