+++ src/pybind/mgr/dashboard/frontend/src/app/ceph/pool/pool-form/pool-form-data.ts
@@ -2,8 +2,8 @@ import { Validators } from '@angular/forms';
 
 import { I18n } from '@ngx-translate/i18n-polyfill';
 
-import { SelectBadgesMessages } from '../../../shared/components/select-badges/select-badges-messages.model';
-import { SelectBadgesOption } from '../../../shared/components/select-badges/select-badges-option.model';
+import { SelectMessages } from '../../../shared/components/select/select-messages.model';
+import { SelectOption } from '../../../shared/components/select/select-option.model';
 import { Pool } from '../pool';
 
 export class PoolFormData {
@@ -17,12 +17,12 @@ export class PoolFormData {
     this.applications = {
       selected: [],
       available: [
-        new SelectBadgesOption(false, 'cephfs', ''),
-        new SelectBadgesOption(false, 'rbd', ''),
-        new SelectBadgesOption(false, 'rgw', '')
+        new SelectOption(false, 'cephfs', ''),
+        new SelectOption(false, 'rbd', ''),
+        new SelectOption(false, 'rgw', '')
       ],
       validators: [Validators.pattern('[A-Za-z0-9_]+'), Validators.maxLength(128)],
-      messages: new SelectBadgesMessages(
+      messages: new SelectMessages(
         {
           empty: i18n('No applications added'),
           selectionLimit: {
