+++ src/pybind/mgr/dashboard/frontend/src/app/shared/components/select-badges/select-badges.component.scss
@@ -1,24 +1,7 @@
 @import '../../../../defaults';
 
-.select-menu-edit {
+::ng-deep .select-menu-edit {
   margin-left: -20px;
-}
-.select-menu-item {
-  display: block;
-  cursor: pointer;
-  border-bottom: 1px solid $color-transparent;
-  font-size: 12px;
-  &:hover {
-    background-color: $color-whitesmoke-gray;
-  }
-}
-.select-menu-item-icon {
-  float: left;
-  padding: 0.5em;
-  width: 3em;
-}
-.select-menu-item-content {
-  padding: 0.5em;
 }
 .badge-remove {
   color: $color-solid-white;
