+++ src/pybind/mgr/dashboard/frontend/src/app/shared/components/select/select-option.model.ts
@@ -1,4 +1,4 @@
-export class SelectBadgesOption {
+export class SelectOption {
   selected: boolean;
   name: string;
   description: string;
